// server/types/api/news.ts
// Generated by https://quicktype.io

export interface INewsResponse {
  data: Data;
}

export interface Data {
  status:       string;
  totalResults: number;
  articles:     Article[];
}

export interface Article {
  source:      Source;
  author:      null | string;
  title:       string;
  description: string;
  url:         string;
  urlToImage:  string;
  publishedAt: string;
  content:     null | string;
}

export interface Source {
  id:   ID;
  name: Name;
}

export enum ID {
  AssociatedPress = "associated-press",
  BbcNews = "bbc-news",
  GoogleNews = "google-news",
  Reuters = "reuters",
  TheNewYorkTimes = "the-new-york-times",
  TheWashingtonPost = "the-washington-post",
}

export enum Name {
  AssociatedPress = "Associated Press",
  BBCNews = "BBC News",
  GoogleNews = "Google News",
  Reuters = "Reuters",
  TheNewYorkTimes = "The New York Times",
  TheWashingtonPost = "The Washington Post",
}