// server/types/api/google.ts
// Generated by https://quicktype.io

export interface IGoogleResponse {
  data: Data;
}

export interface Data {
  kind:              string;
  url:               URL;
  queries:           Queries;
  context:           Context;
  searchInformation: SearchInformation;
  items:             Item[];
}

export interface Context {
  title: string;
}

export interface Item {
  kind:             string;
  title:            string;
  htmlTitle:        string;
  link:             string;
  displayLink:      string;
  snippet:          string;
  htmlSnippet:      string;
  formattedUrl:     string;
  htmlFormattedUrl: string;
  pagemap:          Pagemap;
  cacheId?:         string;
}

export interface Pagemap {
  metatags:       { [key: string]: string }[];
  cse_thumbnail?: CSEThumbnail[];
  website?:       Website[];
  cse_image?:     CSEImage[];
}

export interface CSEImage {
  src: string;
}

export interface CSEThumbnail {
  width:  string;
  height: string;
  src:    string;
}

export interface Website {
  name:        string;
  description: string;
  url:         string;
  image:       string;
}

export interface Queries {
  request: Request[];
}

export interface Request {
  title:          string;
  totalResults:   string;
  searchTerms:    string;
  count:          number;
  startIndex:     number;
  inputEncoding:  string;
  outputEncoding: string;
  safe:           string;
  cx:             string;
  dateRestrict:   string;
}

export interface SearchInformation {
  searchTime:            number;
  formattedSearchTime:   string;
  totalResults:          string;
  formattedTotalResults: string;
}

export interface URL {
  type:     string;
  template: string;
}